+++ test/support/endpoint.ex
@@ -1,0 +1,3 @@
+defmodule Flop.Phoenix.Endpoint do
+  use Phoenix.Endpoint, otp_app: :flop_phoenix
+end