+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.5.0] - 2021-03-23
+
 ### Changed
 
 - Rename `FlopPhoenix` to `Flop.Phoenix`.